+++ packages/ui/src/components/databases/view-fullscreen-button.tsx
@@ -1,25 +1,25 @@
 import { Fullscreen } from 'lucide-react';
 
-import { Link } from '@colanode/ui/components/ui/link';
 import { useDatabase } from '@colanode/ui/contexts/database';
 import { useDatabaseViews } from '@colanode/ui/contexts/database-views';
+import { useNavigation } from '@colanode/ui/contexts/navigation';
 
 export const ViewFullscreenButton = () => {
   const database = useDatabase();
   const views = useDatabaseViews();
+  const navigation = useNavigation();
 
   if (!views.inline) {
     return null;
   }
 
   return (
-    <Link
-      from="/workspace/$userId"
-      to="$nodeId"
-      params={{ nodeId: database.id }}
+    <button
+      type="button"
       className="flex cursor-pointer items-center rounded-md p-1.5 hover:bg-accent"
+      onClick={() => navigation.openNode(database.id, 'database')}
     >
       <Fullscreen className="size-4" />
-    </Link>
+    </button>
   );
 };
